remove ClusterInfo in Serializer and Deserializer. move ByteUtilsTest to common test.
